Bugfix: ProfileDB and SharedFolderDB module can be disable
git-svn-id: file:///home/code/svn/ovd/trunk@5593 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ SessionManager/web/admin/actions.php
@@ -904,11 +904,13 @@ if ($_REQUEST['name'] == 'User') {
 						popup_error(sprintf(_("Unable to delete user '%s'"), $user_login));
 						redirect();
 					}
-					$profiledb = ProfileDB::getInstance();
-					$netfolders = $u->getProfiles();
-					if (is_array($netfolders)) {
-						foreach ($netfolders as $netfolder)
-							$profiledb->remove($netfolder->id);
+					if (Preferences::moduleIsEnabled('ProfileDB')) {
+						$profiledb = ProfileDB::getInstance();
+						$netfolders = $u->getProfiles();
+						if (is_array($netfolders)) {
+							foreach ($netfolders as $netfolder)
+								$profiledb->remove($netfolder->id);
+						}
 					}
 					$res = $userDB->remove($u);
